requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.1.31
eval_type_backport==0.2.2
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jsonlines==4.0.0
mistralai==1.6.0
playwright==1.51.0
pydantic_core==2.27.2
pyee==12.1.1
python-dateutil==2.9.0.post0
python-decouple==3.8
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
typing-inspection==0.4.0
typing_extensions==4.13.0
